refactor(pca): single confirmation-outcome enum + real no-surface test + focused register test file (#1423 R8)
Addresses otReviewAgent round-8 (3 of 4; the 4th contradicts round-7 and is
deferred for a human call):

- 8-B: model the advisory confirmation as a single `PcaConfirmationOutcome`
  discriminant (`confirmed | not_observed | inconclusive | unsupported`) instead
  of two coupled fields. The module-private state machine + the facade method
  return the outcome; the HTTP boundary derives the wire
  `{ verified, adapterSupported }` via a pure exhaustive `pcaConfirmationToWire`.
  Behavior-identical (audited: all outcome paths map to the same wire shapes);
  the state machine collapses to one `sawNotObserved` flag.

- 8-A: the "unsupported" outcome is now tested with a REAL `NoChainAdapter`
  (which lacks the probe method, so the facade's typeof-guard returns the
  no-surface signal) instead of an impossible mock that returns null — the
  scripted null case bypassed the very guard it claimed to cover.

- 8-D: extract the #1085 /register policy-matrix tests + harness out of the
  oversized daemon-http-behavior-extra.test.ts into a focused
  daemon-context-graph-register.test.ts. Bonus: being a pure route-handler test,
  it now runs in the normal unit lane (no daemon-spawn / hardhat harness).

Verified: tsc agent 0 / cli 0; agent pca-v10-facade 30/30 + ensure-registered
7/7; cli daemon-pca-routes 106/106 + daemon-context-graph-register 5/5.
