+++ packages/agent/src/index.ts
@@ -71,7 +71,8 @@ export {
   type PolicyApprovalBinding,
 } from './ccl-policy.js';
 export { DKGAgent } from './dkg-agent.js';
-export type { PcaAgentConfirmation } from './dkg-agent-registry.js';
+export type { PcaAgentConfirmation, PcaConfirmationOutcome } from './dkg-agent-registry.js';
+export { pcaConfirmationToWire } from './dkg-agent-registry.js';
 export {
   verifyBatch,
   buildBatchRejectionRecord,
